pyproject.toml
[project]
name = "currency-converter-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "nikewalce",email = "nikewalce1@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-bdd (>=8.1.0,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[[tool.poetry.packages]]
include = "src_bot"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
